requirements.txt
streamlit>=1.37.0
translate>=3.6.1
python-dotenv>=1.0.1
transformers>=4.41.0
torch>=2.3.0
sentencepiece>=0.2.0
sacremoses>=0.1.1
google-generativeai>=0.5.0
